Cahce files and fix download counter

+++ app/Http/Controllers/DownloadController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Repositories\SaleRepository;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Illuminate\Contracts\Filesystem\FileNotFoundException;
@@ -22,18 +23,19 @@ class DownloadController extends Controller
     public function download(string $id, SaleRepository $saleRepository): BinaryFileResponse
     {
         $attachment = $saleRepository->getDownload($id);
-
-        $fileContents = Storage::disk($attachment->disk)
-            ->get(
-                $attachment->path .
-                $attachment->name .
-                '.' .
-                $attachment->extension
-            );
-
         $tempFile = storage_path('app/tmp/' . $attachment->name);
 
-        file_put_contents($tempFile, $fileContents);
+        if (!File::exists($tempFile)) {
+            $fileContents = Storage::disk($attachment->disk)
+                ->get(
+                    $attachment->path .
+                    $attachment->name .
+                    '.' .
+                    $attachment->extension
+                );
+
+            file_put_contents($tempFile, $fileContents);
+        }
 
         return response()->download(
             $tempFile,
